modify python committed

+++ demo-lambda/dynamodb_delete.py
@@ -4,9 +4,14 @@ dynamodb = boto3.resource("dynamodb")
 def lambda_handler(event, context):
     table = dynamodb.Table("student")
     primary_key = {"id":  event['queryStringParameters']['id']}
-    
-    res = table.delete_item(Key= primary_key)
+    status="200"    
+    if "Item" in table.get_item(Key = primary_key):
+        res = table.delete_item(Key= primary_key)
+        if not res:
+            status = "400" 
+    else:
+        status = "404"
     return {
-    'statusCode': 200,
-    'body': json.dumps(res),
-    }
+        "statusCode": status
+    }   
+
